+++ markdrop/__init__.py
@@ -1,5 +1,9 @@
 from .utils import extract_images, make_markdown, extract_tables_from_pdf
+from .models.img_descriptions import generate_descriptions
+from .models.setup_keys import setup_keys
+from .helper import analyze_pdf_images
 from .ignore_warnings import *
 
-__version__ = "0.1.2"
-__all__ = ['extract_images', 'make_markdown', 'extract_tables_from_pdf']  
+__version__ = "0.2.0"
+__all__ = ['extract_images', 'make_markdown', 'extract_tables_from_pdf', 
+           'generate_descriptions', 'setup_keys', 'analyze_pdf_images']